Integrate linters through pre-commit.com tool (#655)
Integrate linters through pre-commit.com tool

This is a common approach to managing the linting setup in the wider
community that is also already used in some of our repositories.
Some of the checks are disabled and ignored to simplify the initial
adoption but must be addressed later.
P.S. Ignore the failing checks, they will disappear once #654 is merged.

Reviewed-by: Bradley A. Thornton <bthornto@redhat.com>
Reviewed-by: None <None>