zuul: remove dependency on ansible-runner (#675)
zuul: remove dependency on ansible-runner

As we no longer need to perform cross testing of
changes between ansible-runner and ansible-navigator
we remove that dependency.
This greatly improves the performance of zuul testing
and lower the resources usage as we no longer need
to wait for the container to be built.

Reviewed-by: Sviatoslav Sydorenko <webknjaz+github/profile@redhat.com>
Reviewed-by: None <None>