Use a 2-line base indentation style @ tox sections (#681)
Use a 2-line base indentation style @ tox sections

This is a less intrusive version of #677 that implements the same goals. I've skipped touching the non-tox section [flake8] because about to be moved out in a parallel effort. The comment reflow matches most of the other ones but changing long bash commands has also been skipped because editing them is scary (tox tends to break these if they are split incorrectly).

Reviewed-by: Sorin Sbarnea <sorin.sbarnea@gmail.com>
Reviewed-by: None <None>