Update all fixtures (#688)
[tests] Update all fixtures

This is a full update of all fixtures

added a newline to each (this is the main reason for the PR) (see the _common.py file for the code change)
other fixture updates expected:


image name update
uuids
py env name

beacuse some fixtures were missing, but have been added (not required by the specific tests), some new spelling errors have been introduced.  Will address in a different PR
This should limit the changes in #685 while tying the fixture updates to the code change that added the newline.

Reviewed-by: Sorin Sbarnea <sorin.sbarnea@gmail.com>
Reviewed-by: None <None>