Use tox lint job and document it (#708)
Use tox lint job and document it

use 'lint' instead of 'linters' for both tox and zuul jobs, so
they are consistent across all our projects. zuul job name is
already named ansible-navigator-tox-lint so it does not need change.
remove leftover type from envlist in tox as it was removed, as is
part of lint.
remove output lines from lint command that are not only outdated
but also irrelevant for docs inclusion.

Reviewed-by: Bradley A. Thornton <bthornto@redhat.com>
Reviewed-by: Sviatoslav Sydorenko <webknjaz+github/profile@redhat.com>
Reviewed-by: None <None>
Reviewed-by: Sorin Sbarnea <sorin.sbarnea@gmail.com>