look_for in noee config (#731)
Fix and enable tests with py36 and py37

This changes the fialing tests such that they pass on both py36 and py37.
Because an older version of python may pull in an older version of ansible, we cannot directly compare the scrren contents for :config since the config options for different version of ansible may differ.
Only the no ee tests were failing because ansible was running outside the EE. Inside the EE the python version doesn't vary based on the version of python being tested...
This adds "look_fors" for these tests (one near the top and one near the bottom of the output).
Also- renable the py36 and py37 tests so they are voting, and update fixtures to reflect the change from compare_fixture to simply looking for some strings on the screen.
Related: #612

Reviewed-by: Sorin Sbarnea <sorin.sbarnea@gmail.com>
Reviewed-by: None <None>