Make codespell an official linter (#727)
Make codespell an official linter

Make codespell implicit (required) in order to pass reviews.
Errors found by codespell will need to be fixed for PRs to be merged if this is agreed to.
About codespell:
Fix common misspellings in text files. It's designed primarily for checking misspelled words in source code, but it can be used with other files as well. It does not check for word membership in a complete dictionary, but instead looks for a set of common misspellings. Therefore it should catch errors like "adn", but it will not catch "adnasdfasdf". This also means it shouldn't generate false-positives when you use a niche term it doesn't know about.
Our configuration for it is here, it will be run by the precommit hook:
https://github.com/ansible/ansible-navigator/blob/main/.pre-commit-config.yaml#L150
This change is does close a range of preparatory changes that repaired problems identified by code-spell:

#728
#718
#710
#707
#690

Please do not gate this change until the entire team approves it.

Reviewed-by: Sviatoslav Sydorenko <webknjaz+github/profile@redhat.com>
Reviewed-by: None <None>
Reviewed-by: Priyam Sahoo <None>
Reviewed-by: Alison Hart <contact@alisonlhart.com>
Reviewed-by: Jake Jackson  <jljacks93@gmail.com>
Reviewed-by: Bradley A. Thornton <bthornto@redhat.com>
Reviewed-by: Ganesh Nalawade <None>