[lint] Unused var and import error (#751)
[lint] Unused var and import error

Fix an unused variable in actions/config
Use the same approach in 2 other places. (improves the fix in #746)
Disable the import error in utils, since that directory isn't really a package, but instead catalog_collections is just a script that gets run, so a full or relative import couldn't be used in this case.

Reviewed-by: Sorin Sbarnea <sorin.sbarnea@gmail.com>
Reviewed-by: None <None>
Reviewed-by: Sviatoslav Sydorenko <webknjaz+github/profile@redhat.com>
Reviewed-by: Bradley A. Thornton <bthornto@redhat.com>