member-iterable (#754)
[lint] Fix pylint `no-member` and `not-an-iterable`

In the case of fields and options, both are dynamically created via a default factory.
The docs for E1101 state:
False positives: This message may report object members that are created dynamically, but exist at the time they are accessed.
http://pylint-messages.wikidot.com/messages:e1101
Although I could not find a mention for not-an-iterable (E1133), the assumption here is that it is also a false positive

Reviewed-by: None <None>