[configuration] delayed postprocs, extra volmounts (#738)
[configuration] delayed postprocs, extra volmounts

I've pulled this out from my lint work, where it was needed, because it feels like it should be a separate change because we probably want it regardless.
This allows post processors to be delayed until after the first pass of post processors has been completed.
At that point, it will run the delayed ones. There is only one "level" of delay (so delayed post processors can't wait for a third pass). This is intentional, to keep things simple, otherwise we'd have a whole tree of post processors and it gets complex quickly.
Doing it this way just provides a simple way to say "this post processor depends on some other things that happen in the first pass, let's wait for this before we actually run it."
Change:

- Allow post-processors to be delayed until after the first
  (alphabetical) pass. In the delayed pass, delayed post-processors
  are again executed alphabetically.

- In the interest of keeping configuration simple, we don't allow for
  multiple "levels" of delay. A post-processor is either run during the
  first pass, or delayed until after the first/normal pass.

- This also adds an "extra volmounts" variable to the navigator
  post-processor class, which can be used by other post-processors to
  signify that they want a volume to be mounted. The volume-mount post
  processor is now delayed to accommodate.

Test Plan:

- Using this in lint/navigator work, no new tests yet because there is
  no other post processor currently which adds an extra volume mount.

Signed-off-by: Rick Elrod <rick@elrod.me>

Reviewed-by: Bradley A. Thornton <bthornto@redhat.com>
Reviewed-by: Sorin Sbarnea <sorin.sbarnea@gmail.com>
Reviewed-by: Sviatoslav Sydorenko <webknjaz+github/profile@redhat.com>
Reviewed-by: Rick Elrod <rick@elrod.me>
Reviewed-by: None <None>